requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
Django==4.0.4
gitdb==4.0.9
GitPython==3.1.27
idna==3.3
requests==2.27.1
smmap==5.0.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
